# backend/internal/apidocs/openapi.yaml
openapi: 3.1.0
info:
  title: Touchline API
  version: 0.1.0
  description: |
    Touchline is a text-first football management game. The backend is a single
    binary (`cmd/touchline serve` runs the API, the scheduler world clock, and
    the event worker in one process; the api/scheduler/worker subcommands split
    them for prod isolation).

    Authentication is cookie-based: login exchanges credentials for a rotating
    access/refresh cookie pair, and every game endpoint below a security scheme
    marker requires a valid `access_token` cookie. The whole world is scoped to
    the session, so "my world" reads are always the caller's own world.

    Launch model: worlds are created as metadata (name), then an admin declares
    countries and leagues, and finally seeds the whole world with clubs,
    players, and memberships. Seasons/league play follow the fixture system
    driven by the scheduler.

    The interactive reference is this page; the raw spec lives at
    `/openapi.yaml`. Live game state (world ticks, match pace 15-min ticks)
    streams over `/ws` (see the WebSocket section).
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development (cmd/touchline serve)
...
paths:
...
  /api/clubs:
    get:
      tags: [clubs]
      summary: List clubs in my world
      operationId: listClubs
      description: Clubs of the caller's own world (AI clubs included).
      responses:
        "200":
          description: Club list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  clubs:
                    type: array
                    items: { $ref: "#/components/schemas/Club" }
              examples:
                sample:
                  value:
                    clubs:
                      - id: 00000000-0000-4000-8000-000000000005
                        world_id: 00000000-0000-4000-8000-000000000001
                        name: Northgate Rovers
                        short_name: NGR
        "403":
          description: Session has no world context (admin console session).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { error: "no world context" }
...
  /api/admin/club-name-parts:
    get:
      tags: [admin]
      summary: List club name stems and suffixes
      operationId: listClubNameParts
      description: Global pools used by the seeding generator to compose club names.
      responses:
        "200":
          description: The two name-part pools.
          content:
            application/json:
              schema:
                type: object
                properties:
                  stems: { type: array, items: { type: string }, example: [northgate, highfield, eastbrook] }
                  suffixes: { type: array, items: { type: string }, example: [rovers, united, athletic] }
    post:
      tags: [admin]
      summary: Add a club name stem or suffix
      operationId: addClubNamePart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kind, value]
              properties:
                kind: { type: string, enum: [stem, suffix], example: stem }
                value: { type: string, example: silvermoor }
      responses:
        "201":
          description: The upserted part.
          content:
            application/json:
              schema:
                type: object
                properties:
                  kind: { type: string }
                  value: { type: string }
        "400":
          description: Invalid kind or empty value.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/admin/club-name-parts/{kind}/{value}:
    delete:
      tags: [admin]
      summary: Delete a club name stem or suffix
      operationId: removeClubNamePart
      parameters:
        - name: kind
          in: path
          required: true
          schema: { type: string, enum: [stem, suffix] }
        - name: value
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Removed (deleting an absent entry is a no-op).
        "400":
          description: Invalid kind.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: access_token
      description: httpOnly cookie set by POST /api/auth/login and rotated by /api/auth/refresh.
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, example: internal error }
...
    Club:
      type: object
      properties:
        id: { type: string, format: uuid }
        world_id: { type: string, format: uuid }
        name: { type: string }
        short_name: { type: string }
...
security:
  - sessionCookie: []